refactor: turned lexical into a directory
