+++ main.py
@@ -22,10 +22,10 @@ def main():
         # list because I'm using it later, also works as a generator expression
         # we also use this instead of .readlines() because it doesn't give \n
 
-    target = list(parse(text))
-    embedding = list(embed(target))  ## TODO: clusters cannot currently take a map
+    target = parse(text)
+    embedding = list(embed(target))  ## TODO: ward linkage doesn't currently take a map
     clusters = cluster(embedding, target, 1)
-    labels = select(target, clusters)
+    labels = select(text, clusters)
 
     ## TODO: I think clusters should be numeric representation of the cluster that the
     ## position in the original text belongs to rather than the list of clusters
